[Refactor] #158 - Change Type Int 64 to Int

+++ Hankkijogbo/Hankkijogbo/Network/Report/DTO/Response/GetReportedNumberResponseDTO.swift
@@ -10,5 +10,5 @@ import Foundation
 // MARK: - 제보수 조회 API Res
 
 struct GetReportedNumberResponseData: Codable {
-    var count: Int64
+    var count: Int
 }
